requirements.txt
asgiref==3.8.1
Django==5.1
django-cors-headers==4.4.0
django-dotenv==1.4.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.9.0
sqlparse==0.5.1
